# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriSafe IoT Food Spoilage Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SensorReadingInput {
  /** Methane/gas level in ppm */
  methane: number;
  /** Ethanol level in ppm */
  ethanol: number;
  /** Temperature in Celsius */
  temperature: number;
  /** Relative humidity percentage */
  humidity: number;
  /** Device identifier */
  deviceId?: string;
}

export type SensorReadingStatus =
  (typeof SensorReadingStatus)[keyof typeof SensorReadingStatus];

export const SensorReadingStatus = {
  fresh: "fresh",
  warning: "warning",
  spoiled: "spoiled",
} as const;

export interface SensorReading {
  id: number;
  methane: number;
  ethanol: number;
  temperature: number;
  humidity: number;
  status: SensorReadingStatus;
  deviceId?: string | null;
  timestamp: string;
}

export interface Settings {
  id: number;
  /** Methane ppm warning threshold */
  methaneWarning: number;
  /** Methane ppm spoiled threshold */
  methaneSpoiled: number;
  /** Ethanol ppm warning threshold */
  ethanolWarning: number;
  /** Ethanol ppm spoiled threshold */
  ethanolSpoiled: number;
  /** Temperature warning threshold (Celsius) */
  temperatureWarning: number;
  /** Temperature spoiled threshold (Celsius) */
  temperatureSpoiled: number;
  /** Humidity warning threshold (%) */
  humidityWarning: number;
  /** Humidity spoiled threshold (%) */
  humiditySpoiled: number;
}

export interface SettingsInput {
  methaneWarning: number;
  methaneSpoiled: number;
  ethanolWarning: number;
  ethanolSpoiled: number;
  temperatureWarning: number;
  temperatureSpoiled: number;
  humidityWarning: number;
  humiditySpoiled: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetSensorHistoryParams = {
  /**
   * Number of hours of history to fetch (default 24)
   */
  hours?: number;
};