# lib/api-zod/src/generated/types/sensorReading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriSafe IoT Food Spoilage Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { SensorReadingStatus } from "./sensorReadingStatus";

export interface SensorReading {
  id: number;
  methane: number;
  ethanol: number;
  temperature: number;
  humidity: number;
  status: SensorReadingStatus;
  deviceId?: string | null;
  timestamp: Date;
}