# lib/api-zod/src/generated/types/sensorReadingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriSafe IoT Food Spoilage Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface SensorReadingInput {
  /** Methane/gas level in ppm */
  methane: number;
  /** Ethanol level in ppm */
  ethanol: number;
  /** Temperature in Celsius */
  temperature: number;
  /** Relative humidity percentage */
  humidity: number;
  /** Device identifier */
  deviceId?: string;
}