# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriSafe IoT Food Spoilage Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  /** Methane ppm warning threshold */
  methaneWarning: number;
  /** Methane ppm spoiled threshold */
  methaneSpoiled: number;
  /** Ethanol ppm warning threshold */
  ethanolWarning: number;
  /** Ethanol ppm spoiled threshold */
  ethanolSpoiled: number;
  /** Temperature warning threshold (Celsius) */
  temperatureWarning: number;
  /** Temperature spoiled threshold (Celsius) */
  temperatureSpoiled: number;
  /** Humidity warning threshold (%) */
  humidityWarning: number;
  /** Humidity spoiled threshold (%) */
  humiditySpoiled: number;
}