# lib/api-zod/src/generated/types/settingsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriSafe IoT Food Spoilage Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsInput {
  methaneWarning: number;
  methaneSpoiled: number;
  ethanolWarning: number;
  ethanolSpoiled: number;
  temperatureWarning: number;
  temperatureSpoiled: number;
  humidityWarning: number;
  humiditySpoiled: number;
}